[iOS[ fix: RequestBuilder 객체 파일 분리
