+++ iOS/MiniVibe/MiniVibe/Common/Network/NetworkService.swift
@@ -36,40 +36,3 @@ final class NetworkService {
             .eraseToAnyPublisher()
     }
 }
-
-struct RequestBuilder<T: Encodable> {
-    // TODO: - 각 뷰에 맞춰서 urlString을 endpoint에 맞게 생성해주는 객체 생성 필요
-    let urlString: String
-    let method: NetworkMethod = .get
-    let body: T?
-    let headers: [String: String]?
-    
-    func create() -> URLRequest? {
-        guard let url = URL(string: urlString) else { return nil }
-        var request = URLRequest(url: url)
-        request.httpMethod = method.rawValue.uppercased()
-        if let body = body,
-           let data = try? JSONEncoder().encode(body) {
-            request.httpBody = data
-        }
-        if let headers = headers {
-            request.allHTTPHeaderFields = headers
-        }
-        return request
-    }
-}
-
-enum NetworkMethod: String {
-    case get
-    case post
-    case put
-    case patch
-    case delete
-}
-
-struct Welcome: Codable {
-    let userId: Int
-    let id: Int
-    let title: String
-    let completed: Bool
-}
